rei: Changes and jank reduction
- Added an explanitory textured widget to Enlightenment
- Changed Heart Extraction to Vivisection

+++ src/main/java/com/neep/neepmeat/compat/rei/display/EnlighteningDisplay.java
@@ -46,7 +46,7 @@ public class EnlighteningDisplay extends BasicDisplay
         return NMREIPlugin.ENLIGHTENING;
     }
 
-    public float getData()
+    public long getData()
     {
         return recipe.getData();
     }
